修复 eval dtype bug 并补上文件日志
1. eval/evaluate.py: 修复 soft_prefix(float32) 与 input_embeds(bfloat16)
   拼接时的 dtype 不匹配崩溃（RuntimeError: mat1 and mat2 must have the
   same dtype）。rollout.py/trainer.py 里同类拼接早已有 .to(dtype) 转换，
   这里是唯一遗漏的一处，训练时因 eval_freq 与 group_size 不整除从未
   触发过，直到把 eval_freq 改成能整除才第一次暴露。

2. configs/alfworld.yaml: eval_freq 50→40，使其能被 group_size=8 整除
   （之前 episode_count % eval_freq 永远不为 0，eval 从未真正跑过）。

3. train.py: 新增 setup_logging()，读取此前完全未被使用的
   paths.log_dir 配置，把所有模块的日志（通过 root logger 传播）同时
   写入按时间戳命名的文件和终端。

+++ eval/evaluate.py
@@ -112,7 +112,7 @@ def run_eval(
 
                 embed_layer  = model.get_input_embeddings()
                 input_embeds = embed_layer(enc.input_ids)           # [1, seq, H]
-                inputs_embeds = torch.cat([soft_prefix, input_embeds], dim=1)
+                inputs_embeds = torch.cat([soft_prefix.to(input_embeds.dtype), input_embeds], dim=1)
                 prefix_mask   = torch.ones(
                     1, soft_prefix.size(1),
                     dtype=enc.attention_mask.dtype, device=device
